+++ coreax/solvers/__init__.py
@@ -26,6 +26,8 @@ from coreax.solvers.coresubset import (
     ConditionalKernelHerding,
     ConditionalKernelHerdingState,
     HerdingState,
+    JointKernelHerding,
+    JointRPCholesky,
     KernelHerding,
     RandomSample,
     RPCholesky,
@@ -49,4 +51,6 @@ __all__ = [
     "PaddingInvariantSolver",
     "ConditionalKernelHerding",
     "ConditionalKernelHerdingState",
+    "JointKernelHerding",
+    "JointRPCholesky",
 ]