+++ tests/unit/test_solvers.py
@@ -33,17 +33,13 @@ from typing_extensions import override
 
 from coreax.coreset import Coreset, Coresubset
 from coreax.data import Data, SupervisedData
-from coreax.inverses import (
-    LeastSquareApproximator,
-    RandomisedEigendecompositionApproximator,
-)
 from coreax.kernel import (
     Kernel,
     PCIMQKernel,
     SquaredExponentialKernel,
 )
 from coreax.solvers import (
-    ConditionalKernelHerding,
+    JointRPCholesky,
     KernelHerding,
     MapReduce,
     RandomSample,
@@ -57,7 +53,6 @@ from coreax.solvers.base import (
     RefinementSolver,
 )
 from coreax.solvers.coresubset import (
-    ConditionalKernelHerdingState,
     HerdingState,
     RPCholeskyState,
 )
@@ -372,6 +367,50 @@ class TestKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
         assert eqx.tree_equal(state, expected_state)
 
 
+class TestJointKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.JointKernelHerding`."""
+
+    additional_key = jr.key(0)
+
+    @override
+    @pytest.fixture(scope="class")
+    def solver_factory(self) -> jtu.Partial:
+        kernel = SquaredExponentialKernel()
+        coreset_size = self.shape[0] // 10
+        return jtu.Partial(
+            JointKernelHerding,
+            coreset_size=coreset_size,
+            feature_kernel=kernel,
+            response_kernel=kernel,
+        )
+
+    @override
+    @pytest.fixture(params=["random"], scope="class")
+    def reduce_problem(
+        self,
+        request: pytest.FixtureRequest,
+        solver_factory: Union[type[Solver], jtu.Partial],
+    ) -> _ReduceProblem:
+        if request.param == "random":
+            features = jr.uniform(self.random_key, self.shape)
+            responses = jr.uniform(self.additional_key, self.shape)
+            solver = solver_factory()
+            expected_coreset = None
+        else:
+            raise ValueError("Invalid fixture parametrization")
+        return _ReduceProblem(
+            SupervisedData(features, responses), solver, expected_coreset
+        )
+
+    def test_herding_state(self, reduce_problem: _ReduceProblem) -> None:
+        """Check that the cached herding state is as expected."""
+        dataset, solver, _ = reduce_problem
+        solver = cast(JointKernelHerding, solver)
+        _, state = solver.reduce(dataset)
+        expected_state = HerdingState(solver.kernel.gramian_row_mean(dataset))
+        assert eqx.tree_equal(state, expected_state)
+
+
 class TestRandomSample(ExplicitSizeSolverTest):
     """Test cases for :class:`coreax.solvers.coresubset.RandomSample`."""
 
@@ -430,33 +469,22 @@ class TestRPCholesky(ExplicitSizeSolverTest):
         assert eqx.tree_equal(state, expected_state)
 
 
-class TestSteinThinning(RefinementSolverTest, ExplicitSizeSolverTest):
-    """Test cases for :class:`coreax.solvers.coresubset.SteinThinning`."""
+class TestJointRPCholesky(ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.JointRPCholesky`."""
+
+    additional_key = jr.key(0)
 
     @override
     @pytest.fixture(scope="class")
     def solver_factory(self) -> jtu.Partial:
-        kernel = PCIMQKernel()
-        coreset_size = self.shape[0] // 10
-        return jtu.Partial(SteinThinning, coreset_size=coreset_size, kernel=kernel)
-
-
-class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
-    """Test cases for :class:`coreax.solvers.coresubset.ConditionalKernelHerding`."""
-
-    feature_kernel = SquaredExponentialKernel()
-    response_kernel = SquaredExponentialKernel()
-
-    @override
-    @pytest.fixture(scope="class")
-    def solver_factory(self) -> jtu.Partial:
+        kernel = SquaredExponentialKernel()
         coreset_size = self.shape[0] // 10
         return jtu.Partial(
-            ConditionalKernelHerding,
-            random_key=self.random_key,
+            JointRPCholesky,
             coreset_size=coreset_size,
-            feature_kernel=self.feature_kernel,
-            response_kernel=self.response_kernel,
+            random_key=self.random_key,
+            feature_kernel=kernel,
+            response_kernel=kernel,
         )
 
     @override
@@ -467,148 +495,48 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         solver_factory: Union[type[Solver], jtu.Partial],
     ) -> _ReduceProblem:
         if request.param == "random":
-            data_key, supervision_key = jr.split(self.random_key)
-            data = jr.uniform(data_key, self.shape)
-            supervision = jr.uniform(supervision_key, self.shape)
+            features = jr.uniform(self.random_key, self.shape)
+            responses = jr.uniform(self.additional_key, self.shape)
             solver = solver_factory()
             expected_coreset = None
         else:
             raise ValueError("Invalid fixture parametrization")
         return _ReduceProblem(
-            SupervisedData(data=data, supervision=supervision), solver, expected_coreset
+            SupervisedData(features, responses), solver, expected_coreset
         )
 
-    @pytest.fixture(
-        params=[
-            "well-sized",
-            "under-sized",
-            "over-sized",
-            "random-unweighted",
-            "random-weighted",
-        ],
-        scope="class",
-    )
-    def refine_problem(
-        self, request: pytest.FixtureRequest, reduce_problem: _ReduceProblem
-    ) -> _RefineProblem:
-        """
-        Pytest fixture that returns a problem dataset and the expected coreset.
+    @override
+    def check_solution_invariants(
+        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
+    ) -> None:
+        """Check functionality of 'unique' in addition to the default checks."""
+        super().check_solution_invariants(coreset, problem)
+        solver = cast(JointRPCholesky, problem.solver)
+        if solver.unique:
+            _, counts = jnp.unique(coreset.nodes.data, return_counts=True)
+            assert max(counts) <= 1
 
-        An expected coreset of 'None' implies the expected coreset for this solver and
-        dataset combination is unknown.
-
-        We expect the '{well,under,over}-sized' cases to return the same result as a
-        call to 'reduce'. The 'random-unweighted' and 'random-weighted' case we only
-        expect to pass without raising an error.
-        """
-        dataset, solver, expected_coreset = reduce_problem
-        indices_key, weights_key = jr.split(self.random_key)
-        solver = cast(ConditionalKernelHerding, solver)
-        coreset_size = min(len(dataset), solver.coreset_size)
-        # We expect 'refine' to produce the same result as 'reduce' when the initial
-        # coresubset has all its indices equal to negative one.
-        expected_coresubset = None
-        if expected_coreset is None:
-            expected_coresubset, _ = solver.reduce(dataset)
-        elif isinstance(expected_coreset, Coresubset):
-            expected_coresubset = expected_coreset
-        if request.param == "well-sized":
-            indices = Data(-1 * jnp.ones(coreset_size, jnp.int32))
-        elif request.param == "under-sized":
-            indices = Data(-1 * jnp.ones(coreset_size - 1, jnp.int32))
-        elif request.param == "over-sized":
-            indices = Data(-1 * jnp.ones(coreset_size + 1, jnp.int32))
-        elif request.param == "random-unweighted":
-            random_indices = jr.choice(indices_key, len(dataset), (coreset_size,))
-            indices = Data(random_indices)
-            expected_coresubset = None
-        elif request.param == "random-weighted":
-            random_indices = jr.choice(indices_key, len(dataset), (coreset_size,))
-            random_weights = jr.uniform(weights_key, (coreset_size,))
-            indices = Data(random_indices, random_weights)
-            expected_coresubset = None
-        else:
-            raise ValueError("Invalid fixture parametrization")
-        initial_coresubset = Coresubset(indices, dataset)
-        return _RefineProblem(initial_coresubset, solver, expected_coresubset)
-
-    # pylint: disable=duplicate-code
-    def test_greedy_cmmd_state(self, reduce_problem: _ReduceProblem) -> None:
-        """Check that the cached ConditionalKernelHerding state is as expected."""
+    def test_rpcholesky_state(self, reduce_problem: _ReduceProblem) -> None:
+        """Check that the cached RPCholesky state is as expected."""
         dataset, solver, _ = reduce_problem
-        solver = cast(ConditionalKernelHerding, solver)
+        solver = cast(RPCholesky, solver)
         _, state = solver.reduce(dataset)
 
-        x = dataset.data
-        y = dataset.supervision
-        num_data_pairs = x.shape[0]
-
-        feature_gramian = solver.feature_kernel.compute(x, x)
-        response_gramian = solver.response_kernel.compute(y, y)
-
-        if solver.inverse_approximator is None:
-            inverse_approximator = LeastSquareApproximator(self.random_key)
-        else:
-            inverse_approximator = solver.inverse_approximator
-        inverse_feature_gramian = inverse_approximator.approximate(
-            array=feature_gramian,
-            regularisation_parameter=solver.regularisation_parameter,
-            identity=jnp.eye(num_data_pairs),
-        )
-        training_cme = jnp.pad(
-            feature_gramian @ inverse_feature_gramian @ response_gramian,
-            [(0, 1)],
-            mode="constant",
-        )
-        feature_gramian = jnp.pad(feature_gramian, [(0, 1)], mode="constant")
-        response_gramian = jnp.pad(response_gramian, [(0, 1)], mode="constant")
-
-        expected_state = ConditionalKernelHerdingState(
-            feature_gramian=feature_gramian,
-            response_gramian=response_gramian,
-            training_cme=training_cme,
-        )
+        x, y = dataset.data, dataset.supervision
+        gramian_diagonal = jax.vmap(solver.kernel.compute_elementwise)((x, y), (x, y))
+        expected_state = RPCholeskyState(gramian_diagonal)
         assert eqx.tree_equal(state, expected_state)
 
-    # pylint: enable=duplicate-code
 
-    def test_batching(self, reduce_problem: _ReduceProblem) -> None:
-        """Test that batching produces no errors."""
-        dataset, _, _ = reduce_problem
-        solver = ConditionalKernelHerding(
-            random_key=self.random_key,
-            coreset_size=10,
-            feature_kernel=self.feature_kernel,
-            response_kernel=self.response_kernel,
-            batch_size=2,
-        )
-        solver.reduce(dataset)
+class TestSteinThinning(RefinementSolverTest, ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.SteinThinning`."""
 
-    def test_non_uniqueness(self, reduce_problem: _ReduceProblem) -> None:
-        """Test that setting 'unique' to be false produces no errors."""
-        dataset, _, _ = reduce_problem
-        solver = ConditionalKernelHerding(
-            random_key=self.random_key,
-            coreset_size=10,
-            feature_kernel=self.feature_kernel,
-            response_kernel=self.response_kernel,
-            unique=False,
-        )
-        solver.reduce(dataset)
-
-    def test_approximate_inverse(self, reduce_problem: _ReduceProblem) -> None:
-        """Test that using an inverse approximator produces no errors."""
-        dataset, _, _ = reduce_problem
-        solver = ConditionalKernelHerding(
-            random_key=self.random_key,
-            coreset_size=10,
-            feature_kernel=self.feature_kernel,
-            response_kernel=self.response_kernel,
-            inverse_approximator=RandomisedEigendecompositionApproximator(
-                self.random_key
-            ),
-        )
-        solver.reduce(dataset)
+    @override
+    @pytest.fixture(scope="class")
+    def solver_factory(self) -> jtu.Partial:
+        kernel = PCIMQKernel()
+        coreset_size = self.shape[0] // 10
+        return jtu.Partial(SteinThinning, coreset_size=coreset_size, kernel=kernel)
 
 
 class _ExplicitPaddingInvariantSolver(ExplicitSizeSolver, PaddingInvariantSolver):