+++ tests/unit/test_util.py
@@ -26,6 +26,7 @@ import equinox as eqx
 import jax.numpy as jnp
 import numpy as np
 import pytest
+from jax import Array
 from jax.random import key
 from scipy.stats import ortho_group
 
@@ -35,6 +36,7 @@ from coreax.util import (
     difference,
     jit_test,
     pairwise,
+    pairwise_tuple,
     sample_batch_indices,
     solve_qp,
     squared_distance,
@@ -83,6 +85,28 @@ class TestUtil:
         y_array = generator.random((num_points_y, dimension))
         expected_output = np.array([[x - y for y in y_array] for x in x_array])
         output = pairwise(difference)(x_array, y_array)
+        assert jnp.linalg.norm(output - expected_output) == pytest.approx(0.0, abs=1e-3)
+
+    def test_pairwise_tuple(self) -> None:
+        """Test the pairwise_tuple transform on an arbitrary function taking tuples."""
+        num_points = 10
+        dimension = 3
+        gen = np.random.default_rng(1_989)
+        data = jnp.array(gen.random((num_points, dimension)))
+        supervision = jnp.array(gen.random((num_points, dimension)))
+        a = (data, supervision)
+        b = (data[::-1], supervision[::-1])
+
+        def tuple_fn(a: tuple[Array, Array], b: tuple[Array, Array]) -> Array:
+            return jnp.array((a[0] - b[0]) * (a[1] - b[1])).sum()
+
+        expected_output = jnp.zeros((num_points, num_points))
+        for i in range(num_points):
+            for j in range(num_points):
+                expected_output = expected_output.at[i, j].set(
+                    tuple_fn((a[0][i], a[1][i]), (b[0][j], b[1][j]))
+                )
+        output = pairwise_tuple(tuple_fn)(a, b)
         assert jnp.linalg.norm(output - expected_output) == pytest.approx(0.0, abs=1e-3)
 
     @pytest.mark.parametrize(
@@ -169,7 +193,7 @@ class TestUtil:
         # this should error as no sensible result can be found in such a case.
         with pytest.raises(TypeError, match="not a valid JAX array type"):
             solve_qp(
-                kernel_mm="invalid_kernel_mm",
+                kernel_mm="invalid_kernel_mm",  # pyright: ignore
                 gramian_row_mean=np.array([1, 2, 3]),
             )
 
@@ -186,7 +210,7 @@ class TestUtil:
         with pytest.raises(TypeError, match="not a valid JAX array type"):
             solve_qp(
                 kernel_mm=np.array([1, 2, 3]),
-                gramian_row_mean="invalid_gramian_row_mean",
+                gramian_row_mean="invalid_gramian_row_mean",  # pyright: ignore
             )
 
     @pytest.mark.parametrize(